requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
Jinja2==3.1.4
python-docx==1.1.2
itsdangerous==2.2.0
